seng-5709/projects/Kafka/requirements.txt
kafka-python~=2.0.2
requests~=2.26.0
types-requests~=2.25.11
